move server error handling code and BEFORE_APP_SERIALIED hook to entry point setup/ssr
