+++ core-libs/setup/ssr/application/error-handling/server-error.collector.ts
