+++ core-libs/setup/ssr/application/error-handling/server-error.interceptor.ts
