+++ core-libs/setup/ssr/application/index.ts
